stockgenmysql/requirements.txt
mysql.connector

